+++ SyMBac/cell.py
@@ -129,29 +129,31 @@ class Cell:
 
     def create_pm_cell(self):
         if self.is_dividing():
-            new_length = self.length/2
-            daughter_length = self.length - new_length
+            new_length = self.length/2  - self.width/4
+            daughter_length = self.length - new_length  - self.width/4
             self.length = new_length
             self.pinching_sep = 0
             cell_vertices = self.calculate_vertex_list()
             cell_shape = pymunk.Poly(None, cell_vertices)
             self.shape = cell_shape
-            cell_moment = 1
-            cell_mass = 0.0001
+            cell_moment = 100001
+            cell_mass = 1
             cell_body = pymunk.Body(cell_mass,cell_moment)
             cell_shape.body = cell_body
             self.body = cell_body
-            new_x = self.position[0] + self.length/2 * np.cos(self.angle*2)
-            new_y = self.position[1] + self.length/2 * np.sin(self.angle*2)
+            new_x = self.position[0] + (self.length + self.width/2)/2 * np.cos(self.angle*2)
+            new_y = self.position[1] + (self.length+ self.width/2)/2 * np.sin(self.angle*2)
             self.body.position = [new_x, new_y]
             cell_body.angle = self.angle
-            cell_shape.friction=1
+            cell_shape.friction=0
             self.space.add(cell_body, cell_shape)
+            #for x in range(10):
+            #    self.space.step(self.dt)
             daughter_details = {
-                "length": daughter_length*0.9,
+                "length": daughter_length,
                 "width": np.random.normal(self.width_mean,self.width_var),
                 "resolution": self.resolution,
-                "position": [self.position[0] - self.length/2 * np.cos(self.angle*2), self.position[1] - self.length/2 * np.sin(self.angle*2)],
+                "position": [self.position[0] - (self.length+ self.width/2)/2 * np.cos(self.angle*2), self.position[1] - (self.length+ self.width/2)/2 * np.sin(self.angle*2)],
                 "angle": self.angle*np.random.uniform(0.95,1.05),
                 "space": self.space,
                 "dt": self.dt,
@@ -170,15 +172,17 @@ class Cell:
             cell_vertices = self.calculate_vertex_list()
             cell_shape = pymunk.Poly(None, cell_vertices)
             self.shape = cell_shape
-            cell_moment = 1
-            cell_mass = 0.0001
+            cell_moment = 100001
+            cell_mass = 1
             cell_body = pymunk.Body(cell_mass,cell_moment)
             cell_shape.body = cell_body
             self.body = cell_body
             cell_body.position = self.position
             cell_body.angle = self.angle
-            cell_shape.friction=0.01
+            cell_shape.friction=0
             self.space.add(cell_body, cell_shape)
+            #for x in range(10):
+            #    self.space.step(self.dt)
             return cell_body, cell_shape
 
     def is_dividing(self): # This needs to be made constant or a cell can divide in one frame and not another frame
@@ -189,7 +193,7 @@ class Cell:
 
 
     def update_length(self):
-        self.length = self.length + self.growth_rate_constant*self.dt*self.length*np.random.uniform(0.7,1.3)
+        self.length = self.length + self.growth_rate_constant*self.dt*self.length*np.random.uniform(0.5,1.3)
         self.pinching_sep = max(0, self.length - self.max_length + self.width)
         self.pinching_sep = min(self.pinching_sep, self.width - 1)
