+++ SyMBac/phase_contrast_drawing.py
@@ -1,6 +1,5 @@
 import pyglet
 from matplotlib import pyplot as plt
-from pymunk.pyglet_util import DrawOptions
 from skimage.transform import rescale, rotate
 from skimage.util import random_noise
 from joblib import Parallel, delayed
@@ -22,144 +21,10 @@ import copy
 from cupyx.scipy.ndimage import convolve as cuconvolve
 from SyMBac.cell import Cell
 from SyMBac.general_drawing import convolve_rescale, make_images_same_shape, perc_diff, raster_cell, transform_func
-from SyMBac.scene_functions import create_space, step_and_update
+from SyMBac.cell_simulation import create_space, step_and_update
 from SyMBac.trench_geometry import trench_creator
 import psfmodels as psfm
 import cupy as cp
-
-
-def get_trench_segments(space):
-    """    
-    A function which extracts the rigid body trench objects from the pymunk space object. Space object should be passed
-    from the return value of the run_simulation() function
-    
-    Returns
-    -------
-    List of trench segment properties, later used to draw the trench.
-    """
-    trench_shapes = []
-    for shape, body in zip(space.shapes, space.bodies):
-        if body.body_type == 2:
-            trench_shapes.append(shape)
-
-    trench_segment_props = []
-    for x in trench_shapes:
-        trench_segment_props.append([x.bb, x.area, x.a, x.b])
-
-    trench_segment_props = pd.DataFrame(trench_segment_props)
-    trench_segment_props.columns = ["bb", "area", "a", "b"]
-    main_segments = trench_segment_props.sort_values("area", ascending=False).iloc[0:2]
-    return main_segments
-
-def run_simulation(trench_length, trench_width, cell_max_length, cell_width, sim_length, pix_mic_conv, gravity,
-                   phys_iters, max_length_var, width_var, save_dir, lysis_p = 0):
-    """
-    Runs the rigid body simulation of bacterial growth based on a variety of parameters. Opens up a Pyglet window to
-    display the animation in real-time. If the simulation looks bad to your eye, restart the kernel and rerun the
-    simulation. There is currently a bug where if you try to rerun the simulation in the same kernel, it will be
-    extremely slow.
-    
-    Parameters
-    ----------
-
-    trench_length : float
-        Length of a mother machine trench (micron)
-    trench_width : float
-        Width of a mother machine trench (micron)
-    cell_max_length : float
-        Maximum length a cell can reach before dividing (micron)
-    cell_width : float
-        the average cell width in the simmulation (micron)
-    pix_mic_conv : float
-        The micron/pixel size of the image
-    gravity : float
-        Pressure forcing cells into the trench. Typically left at zero, but can be varied if cells start to fall into
-        each other or if the simulation behaves strangely.
-    phys_iters : int
-        Number of physics iterations per simulation frame. Increase to resolve collisions if cells are falling into one
-        another, but decrease if cells begin to repel one another too much (too high a value causes cells to bounce off
-        each other very hard). 20 is a good starting point
-    max_length_var : float
-        Variance of the maximum cell length
-    width_var : float
-        Variance of the maximum cell width
-    save_dir : str
-        Location to save simulation outupt
-    lysis_p : float
-        probability of cell lysis
-        
-    Returns
-    -------
-    cell_timeseries : lists
-        A list of parameters for each cell, such as length, width, position, angle, etc. All used in the drawing of the
-        scene later
-    space : a pymunk space object
-        Contains the rigid body physics objects which are the cells.
-    """
-
-    space = create_space()
-    space.gravity = 0, gravity  # arbitrary units, negative is toward trench pole
-    dt = 1 / 100  # time-step per frame
-    pix_mic_conv = 1 / pix_mic_conv  # micron per pixel
-    scale_factor = pix_mic_conv * 3  # resolution scaling factor
-
-    trench_length = trench_length * scale_factor
-    trench_width = trench_width * scale_factor
-    trench_creator(trench_width, trench_length, (35, 0), space)  # Coordinates of bottom left corner of the trench
-
-    cell1 = Cell(
-        length=cell_max_length * scale_factor,
-        width=cell_width * scale_factor,
-        resolution=60,
-        position=(20 + 35, 10),
-        angle=0.8,
-        space=space,
-        dt=1 / 60,
-        growth_rate_constant=1,
-        max_length=cell_max_length * scale_factor,
-        max_length_mean=cell_max_length * scale_factor,
-        max_length_var=max_length_var * np.sqrt(scale_factor),
-        width_var=width_var * np.sqrt(scale_factor),
-        width_mean=cell_width * scale_factor,
-        parent=None,
-        lysis_p=lysis_p
-    )
-
-
-    window = pyglet.window.Window(700, 700, "SyMBac", resizable=True)
-    options = DrawOptions()
-
-    @window.event
-    def on_draw():
-        window.clear()
-        space.debug_draw(options)
-
-    global cell_timeseries
-    global x
-
-    try:
-        del cell_timeseries
-    except:
-        pass
-    try:
-        del x
-    except:
-        pass
-
-    x = [0]
-    cell_timeseries = []
-    cells = [cell1]
-    pyglet.clock.schedule_interval(step_and_update, interval=dt, cells=cells, space=space, phys_iters=phys_iters,
-                                   ylim=trench_length, cell_timeseries=cell_timeseries, x=x, sim_length=sim_length,
-                                   save_dir=save_dir)
-    pyglet.app.run()
-    # window.close()
-    # phys_iters = phys_iters
-    # for x in tqdm(range(sim_length+250),desc="Simulation Progress"):
-    #    cells = step_and_update(dt=dt, cells=cells, space=space, phys_iters=phys_iters,ylim=trench_length*1.1, cell_timeseries = cell_timeseries, x=x, sim_length = sim_length, save_dir = save_dir)
-    #    if x > 250:
-    #        cell_timeseries.append(deepcopy(cells))
-    return cell_timeseries, space
 
 def generate_PC_OPL(main_segments, offset, scene, mask, media_multiplier, cell_multiplier, device_multiplier,
                     y_border_expansion_coefficient, x_border_expansion_coefficient, fluorescence, defocus):
@@ -533,101 +398,6 @@ def generate_test_comparison(media_multiplier=75, cell_multiplier=1.7, device_mu
         plt.close()
     else:
         return noisy_img, expanded_mask_resized_reshaped.astype(int)
-
-def draw_scene(cell_properties, do_transformation, space_size, offset, label_masks, pinching=True):
-    """
-    Draws a raw scene (no trench) of cells, and returns accompanying masks for training data.
-    
-    Parameters
-    ----------
-    cell properties : list
-        A list of cell properties for that frame
-    do_transformation : bool
-        True if you want cells to be bent, false and cells remain straight as in the simulation
-    space_size : tuple
-        The xy size of the numpy array in which the space is rendered. If too small then cells will not fit. recommend using the get_space_size() function to find the correct space size for your simulation
-    offset : int
-        A necessary parameter which offsets the drawing a number of pixels from the left hand side of the image. 30 is a good number, but if the cells are very thick, then might need increasing. 
-    label_masks : bool
-        If true returns cell masks which are labelled (good for instance segmentation). If false returns binary masks only. I recommend leaving this as True, because you can always binarise the masks later if you want. 
-        
-    Returns
-    -------
-    space, space_masks : 2D numpy array, 2D numpy array
-    
-    space : 2D numpy array
-        Not to be confused with the pyglet object calledspace in some other functions. Simply a 2D numpy array with an image of cells from the input frame properties
-    space_masks : 2D numy array
-        The masks (labelled or bool) for that scene.
-    
-    """
-    space_size = np.array(space_size)  # 1000, 200 a good value
-    space = np.zeros(space_size)
-    space_masks_label = np.zeros(space_size)
-    space_masks_nolabel = np.zeros(space_size)
-    colour_label = [1]
-
-    space_masks = np.zeros(space_size)
-    if label_masks == False:
-        space_masks = space_masks.astype(bool)
-
-    for properties in cell_properties:
-        length, width, angle, position, freq_modif, amp_modif, phase_modif, phase_mult, separation = properties
-        position = np.array(position)
-        x = np.array(position).astype(int)[0] + offset
-        y = np.array(position).astype(int)[1] + offset
-        OPL_cell = raster_cell(length=length, width=width, separation=separation, pinching=pinching)
-
-        if do_transformation:
-            OPL_cell_2 = np.zeros((OPL_cell.shape[0], int(OPL_cell.shape[1] * 2)))
-            midpoint = int(np.median(range(OPL_cell_2.shape[1])))
-            OPL_cell_2[:,
-            midpoint - int(OPL_cell.shape[1] / 2):midpoint - int(OPL_cell.shape[1] / 2) + OPL_cell.shape[1]] = OPL_cell
-            roll_coords = np.array(range(OPL_cell_2.shape[0]))
-            freq_mult = (OPL_cell_2.shape[0])
-            amp_mult = OPL_cell_2.shape[1] / 10
-            sin_transform_cell = transform_func(amp_modif, freq_modif, phase_modif)
-            roll_amounts = sin_transform_cell(roll_coords, amp_mult, freq_mult, phase_mult)
-            for B in roll_coords:
-                OPL_cell_2[B, :] = np.roll(OPL_cell_2[B, :], roll_amounts[B])
-            OPL_cell = (OPL_cell_2)
-
-        rotated_OPL_cell = rotate(OPL_cell, -angle, resize=True, clip=False, preserve_range=True, center=(x, y))
-        cell_y, cell_x = (np.array(rotated_OPL_cell.shape) / 2).astype(int)
-        offset_y = rotated_OPL_cell.shape[0] - space[y - cell_y:y + cell_y, x - cell_x:x + cell_x].shape[0]
-        offset_x = rotated_OPL_cell.shape[1] - space[y - cell_y:y + cell_y, x - cell_x:x + cell_x].shape[1]
-        assert y > cell_y, "Cell has {} negative pixels in y coordinate, try increasing your offset".format(y - cell_y)
-        assert x > cell_x, "Cell has negative pixels in x coordinate, try increasing your offset"
-        space[
-        y - cell_y:y + cell_y + offset_y,
-        x - cell_x:x + cell_x + offset_x
-        ] += rotated_OPL_cell
-
-        def get_mask(label_masks):
-
-            if label_masks:
-                space_masks_label[y - cell_y:y + cell_y + offset_y, x - cell_x:x + cell_x + offset_x] += (
-                                                                                                                 rotated_OPL_cell > 0) * \
-                                                                                                         colour_label[0]
-                colour_label[0] += 1
-                return space_masks_label
-            else:
-                space_masks_nolabel[y - cell_y:y + cell_y + offset_y, x - cell_x:x + cell_x + offset_x] += (
-                                                                                                                   rotated_OPL_cell > 0) * 1
-                return space_masks_nolabel
-                # space_masks = opening(space_masks,np.ones((2,11)))
-
-        label_mask = get_mask(True).astype(int)
-        nolabel_mask = get_mask(False).astype(int)
-        label_mask_fixed = np.where(nolabel_mask > 1, 0, label_mask)
-        mask_borders = find_boundaries(label_mask_fixed, mode="thick", connectivity=2)
-        space_masks = np.where(mask_borders, 0, label_mask_fixed)
-        space_masks = opening(space_masks)
-        if label_masks == False:
-            space_masks = space_masks.astype(bool)
-        space = space * space_masks.astype(bool)
-    return space, space_masks
-
 
 def generate_training_data(interactive_output, sample_amount, randomise_hist_match, randomise_noise_match, sim_length,
                            burn_in, n_samples, save_dir, in_series=False, seed=False):
