+++ setup.py
@@ -16,14 +16,12 @@ setup(
     long_description = (this_directory / "README.md").read_text(), 
     long_description_content_type = 'text/markdown',
     install_requires = [
-        'tensorflow==2.8.0',
-        'elasticdeform==0.4.9', 
-        'tifffile==2021.10.12',
-        'scikit-image==0.18.3' ,
-        'matplotlib==3.4.3',
+        'tifffile',
+        'scikit-image',
+        'matplotlib',
         "tqdm",
         "pandas==1.3.4",
-        "natsort==7.1.1",
+        "natsort",
         "jupyterlab",
         "ipywidgets",
         "joblib",
